+++ CelestiumPT/src/acceleration_structure/BVHNode.cuh
@@ -7,7 +7,7 @@ struct BVHNode {
 	BVHNode() = default;
 
 	Bounds3f m_BoundingBox;
-	int left_child_or_triangle_indices_start_idx = -1;	//TODO:make them uint?
+	long int left_child_or_triangle_indices_start_idx = -1;	//TODO:make them uint?
 	int triangle_indices_count = 0;
 
 	float getSurfaceArea() const
